pyproject.toml
[project]
name = "booking-sync"
version = "0.1.0"
description = "Booking system sync backend"
authors = [
    {name = "Sadman Naib", email = "sadman.naib@gmail.com"},
]
dependencies = [
    "django>=6.0.3",
    "djangorestframework>=3.16.1",
    "requests>=2.32.5",
    "psycopg[binary]>=3.3.3",
    "celery>=5.6.2",
    "redis>=7.3.0",
    "django-celery-beat>=2.9.0",
    "python-dotenv>=1.2.2",
    "dj-database-url>=3.1.2",
]
requires-python = "==3.12.*"
readme = "README.md"
license = {text = "None"}

[tool.pdm]
distribution = false

[tool.pdm.dev-dependencies]
dev = [
    "black>=26.3.0",
    "isort>=8.0.1",
    "pytest>=9.0.2",
    "pytest-django>=4.12.0",
]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings"
testpaths = ["tests"]
python_files = "test_*.py"
addopts = "--verbose"
